Code Refactor
When I used to write code, I for some reason thought it good practice to include
loads of whitespace.

Most of the things I have changed are removing said whitespace and including
list comprehensions.

Next, I would like to use dataclasses for some of the classes.
Futhermore be able to use protocols for RK method as is done in swift.
Implementation would look slightly different but the principles are the same.